Refactor snap parsing into standalone functions
